Begin work on the skills data
