+++ web/public/locales/fr/form.json
@@ -6,5 +6,12 @@
     "courses": "Cours de 10 jours et de 3 jours",
     "dayZero": "Jour 0",
     "remote": "À distance"
+  },
+  "categories": {
+    "label": "Intérêts",
+    "kitchen": "Cuisine",
+    "maintenance": "Construction et maintenance",
+    "technology": "Technologie et design",
+    "accounting": "Comptabilité"
   }
 }
