+++ web/src/components/FormPageSkills/FormPageSkills.js
@@ -1,21 +1,13 @@
 import * as yup from 'yup'
-import { Label, Checkbox } from 'theme-ui'
 import isEqual from 'lodash/isEqual'
 import isEmpty from 'lodash/isEmpty'
-import { useFormContext, Controller } from 'react-hook-form'
-import { motion } from 'framer-motion'
 
 import FormPageLayout from 'src/layouts/FormPageLayout/FormPageLayout'
 import { useFormService } from 'src/hooks/useFormService'
-import {
-  Form,
-  FormField,
-  FormInputChoice,
-  FormInputChoiceBox,
-} from 'src/components/FormComponents/FormComponents'
-
-import FormPageAnimation from '../FormPageAnimation/FormPageAnimation'
-import FormButtonWrapper from '../FormButtonWrapper/FormButtonWrapper'
+import { Form } from 'src/components/FormComponents/FormComponents'
+import FormPageAnimation from 'src/components/FormPageAnimation/FormPageAnimation'
+import FormButtonWrapper from 'src/components/FormButtonWrapper/FormButtonWrapper'
+import SkillsCell from 'src/components/SkillsCell/SkillsCell'
 
 export const skillSchema = yup.object().shape({
   _hidden: yup.mixed().required(),
@@ -108,121 +100,13 @@ const FormPageSkills = () => {
         validationResolver={skillResolver}
       >
         <FormPageAnimation motionKey="skills">
-          <FormInputChoice
-            type="checkbox"
-            label="Areas of interest"
-            name="categories"
-            options={[
-              { label: 'Kitchen', value: 'kitchen' },
-              { label: 'Maintenance', value: 'maintenance' },
-              { label: 'Technology', value: 'technology' },
-              { label: 'Accounting', value: 'accounting' },
-            ]}
-          />
-
-          <ConditionalCheckboxes />
-
-          <FormField label="Other skills" name="other" as="textarea" />
-
-          <FormInputChoice
-            type="checkbox"
-            name="consent"
-            options={[
-              {
-                label:
-                  'I consent to being contacted by committees for recruitment purposes',
-              },
-            ]}
-          />
+          <SkillsCell />
         </FormPageAnimation>
 
         <FormButtonWrapper />
       </Form>
     </FormPageLayout>
   )
-}
-
-const skillsData = {
-  kitchen: [
-    'question 1',
-    'question 2',
-    'question 3',
-    'question 4',
-    'question 5',
-    'question 6',
-    'question 7',
-    'question 8',
-  ],
-  maintenance: [
-    'painting',
-    'plumbing',
-    'carpentry',
-    'electricity',
-    'architecture',
-    'etc',
-  ],
-  technology: [
-    'sysadmin',
-    'software development',
-    'tech support',
-    'networking',
-    'etc',
-  ],
-  accounting: ['accounting', 'bookkeeping', 'etc'],
-}
-
-/**
- * Skill selection checkboxes that appear when a category is selected
- * A `react-hook-form` Controller is used to register the checkbox group
- * as the component mounts. The checkboxes themselves are registered directly.
- */
-const ConditionalCheckboxes = () => {
-  const { watch, control, register } = useFormContext()
-
-  // when a category is selected, the component is rerendered
-  // so we have an updated array of user-selected categories
-  const selectedCategories = watch('categories')
-
-  // skillsData is then rendered as a checkbox group for each category
-  if (selectedCategories) {
-    return (
-      <>
-        {selectedCategories.map((category) => (
-          <Controller
-            as={FormInputChoiceBox}
-            key={category}
-            label={category}
-            control={control}
-            // this template string creates a skill object with each
-            // selected category as properties when the form is submitted
-            name={`skills.${category}`}
-            animate
-          >
-            {skillsData[category].map((skill) => (
-              <motion.div
-                variants={{
-                  hidden: { opacity: 0 },
-                  show: { opacity: 1 },
-                }}
-                key={skill}
-              >
-                <Label>
-                  <Checkbox
-                    // and a skill property under the category
-                    name={`skills[${category}][${skill}]`}
-                    ref={register}
-                  />
-                  {skill}
-                </Label>
-              </motion.div>
-            ))}
-          </Controller>
-        ))}
-      </>
-    )
-  }
-
-  return null
 }
 
 export default FormPageSkills
